refactor: create Decaf.Client.Internal.Apis module
... and move Barista, Microlot and Pdms module under this new module.
